minor optimization
prior to this update, the behavior of the winners decision was to iterate over
the rows checking for a full set and then to iterate over the columns for the same.

each of those loops would break as soon as a winner was found.

but if a winner was found in a row, the loop over the columns would still run, and
even potentially add the board to the winners list twice.

when the .each method completes, it returns the array it looped over.  however,
if break runs, the return from the loop is break's argument (nil, in this case).

so i can test for this and run next when a winner is found to move on to the
next board without double-checking the columns.

--- a/lib/aoc2021/day4.rb
+++ b/lib/aoc2021/day4.rb
@@ -19,7 +19,7 @@ module Aoc2021
     def get_winners
       winners = []
       @boards.each do |board|
-        board.each do |row|
+        next unless board.each do |row|
           winners << board and break if row.count { |x| x[:drawn] } == @board_size
         end
         0.upto(@board_size - 1) do |col_index|