2021 day 7
showerthought: the fuel costs form a well, where once you find the
bottom of the well, you can stop searching because all of the costs
after that will continue to increase.

this also means that searching linearly like this is sub-optimal on
a large data set and it'd be faster to binary-search the space.